+++ protocols/ping/CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.43.0 - unreleased
+
+- Remove deprecated items. See [PR 3702].
+  * `Ping` prefixed items.
+  * Keep Alive option 
+
+[PR 3702]: https://github.com/libp2p/rust-libp2p/pull/3702
+
 # 0.42.0
 
 - Update to `libp2p-core` `v0.39.0`.